libjulia_jll 1.7 (#3214)
* libjulia_jll 1.7

* Change most libjulia dependencies into build dependencies

+++ L/libjulia/common.jl
@@ -2,6 +2,8 @@
 # `julia build_tarballs.jl --help` to see a usage message.
 using BinaryBuilder, Pkg
 
+include("../../fancy_toys.jl") # for get_addable_spec
+
 # return the platforms supported by libjulia
 function libjulia_platforms(julia_version)
     platforms = supported_platforms(; experimental=julia_version ≥ v"1.7")
@@ -27,6 +29,7 @@ function build_julia(ARGS, version)
         v"1.4.2" => "76a94e06e68fb99822e0876a37c2ed3873e9061e895ab826fd8c9fc7e2f52795",
         v"1.5.3" => "be19630383047783d6f314ebe0bf5e3f95f82b0c203606ec636dced405aab1fe",
         v"1.6.0" => "c253360d29abb9a3a9e6e01493e0f7bb537d88014fd58ac561b5ba30fcb44cad",
+        v"1.7.0-beta2" => "7bd9ae5b20f9a03ef97f830285007eaa318a780675ab59b6528cd8ef242aa8cb",
     )
 
     sources = [
@@ -77,10 +80,13 @@ function build_julia(ARGS, version)
     override OS=Linux
     EOM
 
+    # TODO: eventually we should get LLVM_CXXFLAGS from llvm-config from
+    # a HostDependency, now that we have those
+    LLVM_CXXFLAGS="-I${prefix}/include -fno-exceptions -fno-rtti -D_GNU_SOURCE -D__STDC_CONSTANT_MACROS -D__STDC_FORMAT_MACROS -D__STDC_LIMIT_MACROS"
     if [[ "${version}" == 1.[0-5].* ]]; then
-        LLVM_CXXFLAGS="-I${prefix}/include -std=c++11 -fno-exceptions -fno-rtti -D_GNU_SOURCE -D__STDC_CONSTANT_MACROS -D__STDC_FORMAT_MACROS -D__STDC_LIMIT_MACROS"
+        LLVM_CXXFLAGS="${LLVM_CXXFLAGS} -std=c++11""
     else
-        LLVM_CXXFLAGS="-I${prefix}/include -std=c++14 -fno-exceptions -fno-rtti -D_GNU_SOURCE -D__STDC_CONSTANT_MACROS -D__STDC_FORMAT_MACROS -D__STDC_LIMIT_MACROS"
+        LLVM_CXXFLAGS="${LLVM_CXXFLAGS} -std=c++14""
     fi
     LLVM_LDFLAGS="-L${prefix}/lib"
     LDFLAGS="-L${prefix}/lib"
@@ -100,8 +106,13 @@ function build_julia(ARGS, version)
             LLVMLINK="-L${prefix}/lib -lLLVM-8jl"
         elif [[ "${version}" == 1.5.* ]]; then
             LLVMLINK="-L${prefix}/lib -lLLVM-9jl"
-        else
+        elif [[ "${version}" == 1.6.* ]]; then
             LLVMLINK="-L${prefix}/lib -lLLVM-11jl"
+        elif [[ "${version}" == 1.7.* ]]; then
+            LLVMLINK="-L${prefix}/lib -lLLVM-12jl"
+        else
+            echo "Error, LLVM version not specified"
+            exit 1
         fi
     fi
 
@@ -170,7 +181,7 @@ function build_julia(ARGS, version)
         USE_SYSTEM_LAPACK=1
         LIBLAPACKNAME=libopenblas
     EOM
-    else
+    elif [[ "${version}" == 1.[0-6].* ]]; then
         if [[ "${version}" == 1.[0-4].* ]]; then
             # remove broken dylib placeholder to force static linking
             rm -f ${prefix}/lib/libosxunwind.dylib
@@ -229,6 +240,10 @@ function build_julia(ARGS, version)
     install_license LICENSE.md
     """
 
+    # HACK: JLLs are not allowed to use prerelease versions, so strip that out
+    # e.g. for 1.7.0-beta2
+    version = VersionNumber(version.major, version.minor, version.patch)
+
     # These are the platforms we will build for by default, unless further
     # platforms are passed in on the command line
     platforms = libjulia_platforms(version)
@@ -251,51 +266,73 @@ function build_julia(ARGS, version)
     ]
 
     # Dependencies that must be installed before this package can be built/used
-    dependencies = [
+
+    dependencies = BinaryBuilder.AbstractDependency[
         Dependency("LibUnwind_jll"),
-        Dependency("PCRE2_jll", compat="10.31"),
-        Dependency("OpenLibm_jll"),
-        Dependency("dSFMT_jll"),
-        Dependency("LibUV_jll"),
-        Dependency("utf8proc_jll"),
-        Dependency("MbedTLS_jll"),
-        Dependency("LibSSH2_jll"),
-        Dependency("LibCURL_jll"),
-        Dependency("Zlib_jll"),
-        Dependency("p7zip_jll"),
-        Dependency("MPFR_jll"),
-        Dependency("GMP_jll"),
-        Dependency("Objconv_jll"),
+        BuildDependency("OpenLibm_jll"),
+        BuildDependency("dSFMT_jll"),
+        BuildDependency("LibUV_jll"),
+        BuildDependency("utf8proc_jll"),
+        BuildDependency("MbedTLS_jll"),
+        BuildDependency("LibSSH2_jll"),
+        BuildDependency("LibCURL_jll"),
+        BuildDependency("Zlib_jll"),
+        BuildDependency("p7zip_jll"),
+        BuildDependency("MPFR_jll"),
+        BuildDependency("GMP_jll"),
+        BuildDependency("Objconv_jll"),
     ]
     if version < v"1.5.1"
         push!(dependencies, Dependency("LibOSXUnwind_jll", compat="0.0.5"))
-    else
+    elseif version < v"1.7"
         push!(dependencies, Dependency("LibOSXUnwind_jll", compat="0.0.6"))
     end
 
     if version < v"1.6"
-        push!(dependencies, Dependency("SuiteSparse_jll", compat="5.4.0"))
+        push!(dependencies, BuildDependency("SuiteSparse_jll", compat="5.4.0"))
     else
-        push!(dependencies, Dependency("SuiteSparse_jll"))
+        push!(dependencies, BuildDependency("SuiteSparse_jll"))
+    end
+
+    if version < v"1.7"
+        push!(dependencies, BuildDependency("PCRE2_jll", compat="10.31"))
+    #else
+    #    push!(dependencies, BuildDependency("PCRE2_jll", compat="10.36"))
     end
 
     if version.major == 1 && version.minor == 3
-        push!(dependencies, Dependency("OpenBLAS_jll", compat="0.3.5"))
+        push!(dependencies, BuildDependency("OpenBLAS_jll", compat="0.3.5"))
         # there is no libLLVM_jll 6.0.1, so we use LLVM_jll instead
         push!(dependencies, Dependency("LLVM_jll", compat="6.0.1"))
-        push!(dependencies, Dependency("LibGit2_jll", compat="0.28.2"))
+        push!(dependencies, BuildDependency("LibGit2_jll", compat="0.28.2"))
     elseif version.major == 1 && version.minor == 4
-        push!(dependencies, Dependency("OpenBLAS_jll", compat="0.3.5"))
+        push!(dependencies, BuildDependency("OpenBLAS_jll", compat="0.3.5"))
         push!(dependencies, Dependency("libLLVM_jll", compat="8.0.1"))
-        push!(dependencies, Dependency("LibGit2_jll", compat="0.28.2"))
+        push!(dependencies, BuildDependency("LibGit2_jll", compat="0.28.2"))
     elseif version.major == 1 && version.minor == 5
-        push!(dependencies, Dependency("OpenBLAS_jll", compat="0.3.9"))
+        push!(dependencies, BuildDependency("OpenBLAS_jll", compat="0.3.9"))
         push!(dependencies, Dependency("libLLVM_jll", compat="9.0.1"))
-        push!(dependencies, Dependency("LibGit2_jll", compat="0.28.2"))
+        push!(dependencies, BuildDependency("LibGit2_jll", compat="0.28.2"))
     elseif version.major == 1 && version.minor == 6
-        push!(dependencies, Dependency("OpenBLAS_jll", compat="0.3.10"))
+        push!(dependencies, BuildDependency("OpenBLAS_jll", compat="0.3.10"))
         push!(dependencies, Dependency("libLLVM_jll", compat="11.0.0"))
-        push!(dependencies, Dependency("LibGit2_jll", compat="1.0.1"))
+        push!(dependencies, BuildDependency("LibGit2_jll", compat="1.0.1"))
+    elseif version.major == 1 && version.minor == 7
+        #push!(dependencies, BuildDependency("OpenBLAS_jll", compat="0.3.13"))
+        #push!(dependencies, Dependency("libLLVM_jll", compat="12.0.0"))
+        #push!(dependencies, BuildDependency("LibGit2_jll", compat="1.0.1"))
+
+        # HACK: we can't install LLVM 12 JLLs for Julia 1.7 from within Julia 1.6. Similar
+        # for several other standard JLLs.
+        # So we use get_addable_spec below to "fake it" for now.
+        # This means the resulting package has fewer dependencies declared, but at least it
+        # will work and allow people to build JLL binaries ready for Julia 1.7
+        push!(dependencies, BuildDependency(get_addable_spec("LLVM_full_jll", v"12.0.0+0")))
+
+        # starting with Julia 1.7, we need LLVMLibUnwind_jll
+        push!(dependencies, BuildDependency(get_addable_spec("LLVMLibUnwind_jll", v"11.0.1+1")))
+    else
+        error("Unsupported Julia version")
     end
 
     julia_compat = version ≥ v"1.7" ? "1.6" : "1.0"
